requirements.txt
defusedxml==0.5.0
docopt==0.6.2
et-xmlfile==1.0.1
get==2018.11.19
jdcal==1.4
mysql==0.0.2
mysql-connector-python==8.0.13
mysqlclient==1.3.14
odfpy==1.4.0
openpyxl==2.5.12
post==2018.11.20
public==2018.11.20
PyYAML==3.13
query-string==2018.11.20
records==0.5.2
request==2018.11.20
SQLAlchemy==1.2.14
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
